Refactor : move method
